pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.0",
]
